test: rename files and remove some unused deps
