+++ package.json
@@ -39,6 +39,7 @@
   "devDependencies": {
     "aegir": "^17.0.0",
     "chai": "^4.2.0",
+    "detect-node": "^2.0.4",
     "dirty-chai": "^2.0.1",
     "ipfs-unixfs-exporter": "~0.35.4",
     "ipld": "~0.20.0",
