+++ test/builder-balanced.spec.js
