+++ test/builder-dir-sharding.spec.js
