+++ test/builder-only-hash.spec.js
