+++ test/hash-parity-with-go-ipfs.spec.js
@@ -1,7 +1,7 @@
 /* eslint-env mocha */
 'use strict'
 
-const importer = require('./../src')
+const importer = require('../src')
 
 const chai = require('chai')
 chai.use(require('dirty-chai'))
